refactor(browser): centralize upload path confinement
